+++ Exercism/c/gigasecond/src/gigasecond.c
@@ -1,0 +1,6 @@
+#include <stdio.h>
+#include "gigasecond.h"
+
+time_t gigasecond_after(time_t start) {
+    return start + 1e9;
+}